feat: move fnd interface
