+++ packages/interface/src/api/funding/apiFnd001.ts
