+++ packages/interface/src/api/funding/apiFnd002.ts
