+++ packages/interface/src/api/funding/apiFnd003.ts
