+++ packages/interface/src/api/funding/apiFnd006.ts
