--- packages/interface/src/api/funding/index.ts
@@ -1,6 +1,0 @@
-import "./endpoint/apiFnd001";
-import "./endpoint/apiFnd002";
-import "./endpoint/apiFnd003";
-import "./endpoint/apiFnd004";
-import "./endpoint/apiFnd005";
-import "./endpoint/apiFnd006";
